+++ R/mx_crosscheck.R
@@ -9,15 +9,35 @@ mx_crosscheck <- function(){
 
   mx_info()
 
-  page <- xml2::read_html("https://www.medrxiv.org/search/%252A")
 
-  reference <- page %>%
-    rvest::html_nodes("#page-title") %>%
+# Get number of unique records in the medRxiv archive ---------------------
+
+  page <- xml2::read_html("https://www.medrxiv.org/archive")
+
+  page_no <- page %>%
+    rvest::html_nodes(".pager-last a") %>%
     rvest::html_text()
 
-  reference <- gsub(",","",reference)
+  page_no <- as.numeric(page_no)-1 # Important as offset by one!
 
-  reference <- as.numeric(stringr::word(reference))
+  page <-
+    xml2::read_html(
+      paste0(
+        "https://www.medrxiv.org/",
+        "archive?field_highwire_a_epubdate_value%5Bvalue%5D&page=",
+        page_no
+      )
+    )
+
+  tmp <- page %>%
+    rvest::html_nodes(".highwire-cite-linked-title") %>%
+    rvest::html_attr('href') %>%
+    data.frame(stringsAsFactors = FALSE)
+
+  reference <- length(tmp$.) + (page_no)*10
+
+
+# Get number of unique records extracted ----------------------------------
 
   data <- suppressMessages(mx_search("*"))